import cycle. wip moved metadata into component package.
