+++ cardinal/ecs/world.go
@@ -40,8 +40,8 @@ type World struct {
 	systemLoggers            []*ecslog.Logger
 	systemNames              []string
 	tick                     uint64
-	nameToComponent          map[string]IComponentMetaData
-	registeredComponents     []IComponentMetaData
+	nameToComponent          map[string]component.IComponentMetaData
+	registeredComponents     []component.IComponentMetaData
 	registeredTransactions   []transaction.ITransaction
 	registeredReads          []IRead
 	isComponentsRegistered   bool
@@ -119,7 +119,7 @@ func RegisterComponent[T component.Component](world *World) error {
 	if err == nil {
 		return fmt.Errorf("component with name '%s' is already registered", t.Name())
 	}
-	c := NewComponentMetaData[T]()
+	c := component.NewComponentMetaData[T]()
 	err = c.SetID(world.nextComponentID)
 	if err != nil {
 		return err
@@ -131,7 +131,7 @@ func RegisterComponent[T component.Component](world *World) error {
 	return nil
 }
 
-func (w *World) GetComponentByName(name string) (IComponentMetaData, error) {
+func (w *World) GetComponentByName(name string) (component.IComponentMetaData, error) {
 	componentType, exists := w.nameToComponent[name]
 	if !exists {
 		return nil, fmt.Errorf("Component with name %s not found. Must register component before using", name)
@@ -212,7 +212,7 @@ func NewWorld(s storage.WorldStorage, sm store.IManager, opts ...Option) (*World
 		namespace:         "world",
 		tick:              0,
 		systems:           make([]System, 0),
-		nameToComponent:   make(map[string]IComponentMetaData),
+		nameToComponent:   make(map[string]component.IComponentMetaData),
 		txQueue:           transaction.NewTxQueue(),
 		Logger:            logger,
 		isGameLoopRunning: atomic.Bool{},
@@ -453,7 +453,7 @@ func (w *World) LoadGameState() error {
 	return nil
 }
 
-func (w *World) loadFromKey(key string, cm storage.ComponentMarshaler, comps []IComponentMetaData) error {
+func (w *World) loadFromKey(key string, cm storage.ComponentMarshaler, comps []component.IComponentMetaData) error {
 	buf, ok, err := w.store.StateStore.Load(key)
 	if !ok {
 		// There is no saved data for this key
@@ -468,7 +468,7 @@ func (w *World) nextEntity() (entity.ID, error) {
 	return w.store.EntityMgr.NewEntity()
 }
 
-func (w *World) insertArchetype(comps []IComponentMetaData) archetype.ID {
+func (w *World) insertArchetype(comps []component.IComponentMetaData) archetype.ID {
 	w.store.ArchCompIdxStore.Push(comps)
 	archID := archetype.ID(w.store.ArchAccessor.Count())
 
@@ -617,7 +617,7 @@ func (w *World) GetTransactionReceiptsForTick(tick uint64) ([]receipt.Receipt, e
 	return w.receiptHistory.GetReceiptsForTick(tick)
 }
 
-func (w *World) GetComponents() []IComponentMetaData {
+func (w *World) GetComponents() []component.IComponentMetaData {
 	return w.registeredComponents
 }
